standalone docs and module structure
Splits src/cli.rs (1.4k lines) into src/cli/{mod,pull,read,call}.rs as
pure moves (clap types/dispatch/run/write_json/open_store_checked in
mod.rs; pull/do_pull/PullError/current-file helpers in pull.rs; the 16
read-only query commands in read.rs; tools/call/import-variables in
call.rs), tightening pub(crate) to pub(super) wherever a helper is only
needed inside the cli subtree. Adds docs/SPEC.md, a consolidated
current-state spec verified against src/ (architecture, data model,
pipeline/sinks, sync/sweep, serve multi-file/watch/MCP, the cached
proxy, variables, CLI surface, determinism contracts) superseding
docs/history/*'s version archaeology. Rewrites README.md per the
standalone-repo spec: install from GitHub Releases (tarball, chmod,
macOS quarantine note, unsigned-binary limitation) plus build-from-
source, the fold-license sentence, corrected `-p figmog`-free command
examples, and limitations reorganized around variables/desktop-proxy/
polling/single-writer. Adds CLAUDE.md (determinism, frozen schema, fold
never vendored, gates, no new deps without justification, synthetic
fixtures, plus a fold-pin-bump checklist entry and the empty
[workspace] table's rationale, also commented in Cargo.toml itself).

Also folds in the deferred debt-ledger minors this same restructuring
touches: a cache-write failure now logs to stderr and still returns the
successful upstream result instead of failing the call (proxy.rs);
PullError moves to a thiserror derive; `figmog tools` no longer opens
the store or requires a resolved mirror (dispatched before resolve_db,
like serve); output-shape e2e tests for `tools`/`import-variables`;
`pub(crate) mod serve` drops the redundant pub(crate) (already
crate-wide from the root); open_store_checked's doc comment gains a
parallel-test panic-hook race note.

138 -> 142 tests. cargo test/clippy --all-targets -D warnings/fmt
--check/doc --no-deps all green on both feature configs.
