+++ CLAUDE.md
@@ -1,0 +1,47 @@
+# Working in this repo
+
+figmog is a fold-backed local mirror of one or more Figma files: a CLI
+(`pull` + read commands) and an MCP stdio server (`serve`) sharing one
+engine. See [`docs/SPEC.md`](docs/SPEC.md) for the current-state spec.
+`docs/history/` holds the retired build-design specs and plans
+(bench/REPL/`--interactive`/`figmog watch`/human-mode output — all
+removed) verbatim, for provenance only — don't treat them as current.
+
+## Standing rules
+
+- **Determinism.** Sorted output at every boundary; no `HashMap`
+  iteration ever reaches an output boundary; `serde_json` is never used
+  with the `preserve_order` feature.
+- **On-disk schema is frozen.** Pipeline sink names (`nodes`, `children`,
+  `text`, `instances_of`, `styled_by`, `bound_to`, `by_type`,
+  `components`, `component_sets`, `styles`, `variables`,
+  `variable_collections`, `meta`, `proxy_cache`) are on-disk schema —
+  renaming one changes what store directory an unmodified binary can
+  open. `Id`/`Rec` are append-only enums: postcard encodes variant
+  indices positionally, so a new variant goes at the end of each enum,
+  never inserted or reordered.
+- **`fold`/`bogkit` is upstream.** Never vendored, never patched — consume
+  only the pinned git dependency's public API.
+- **Gates for any change:** `cargo test`, `cargo clippy --no-deps
+  --all-targets -- -D warnings`, `cargo fmt --check`, and `cargo doc
+  --no-deps` warning-free.
+- **No new dependencies without a written justification in the PR.**
+- **Fixtures are synthetic only** — nothing derived from real client
+  files.
+- **Pin-bump checklist.** Bumping `fold`'s pinned `rev` in `Cargo.toml`
+  isn't just a version bump: `cli/mod.rs`'s `open_store_checked` matches
+  the substring `"Locked"` against a caught panic's payload to distinguish
+  fjall's lock-contention panic from every other kind (see its doc
+  comment). Re-verify that substring still appears in the panic message a
+  locked-store `fold`/fjall open actually produces on any pin bump — a
+  wording change upstream would silently turn every "store is locked"
+  case into `internal panic: ...` + exit 101 instead of the intended
+  clean exit-1 message.
+- **The empty `[workspace]` table in `Cargo.toml` is deliberate, not
+  leftover.** It guards against Cargo silently treating figmog as a
+  member of some *enclosing* workspace if this repo is ever checked out
+  nested inside another Cargo project — an empty `[workspace]` pins this
+  crate as its own workspace root regardless of what's above it.
+  Harmless (and currently a no-op) as a standalone repo with nothing
+  above it; kept for that scenario anyway, and commented in place in
+  `Cargo.toml` itself.
